Move Redis jobs to new location.
Jobs are independent of the rest of funflow, so it makes sense to move them
to their own package. We can then add other ways to distribute jobs to this
same package.
