+++ funflow-jobs/ChangeLog.md
@@ -1,0 +1,5 @@
+# Revision history for funflow-jobs
+
+## 0.1.0.0  -- YYYY-mm-dd
+
+* First version. Released on an unsuspecting world.
