+++ funflow-jobs/funflow-jobs.cabal
@@ -1,0 +1,36 @@
+-- Initial funflow-jobs.cabal generated by cabal init.  For further 
+-- documentation, see http://haskell.org/cabal/users-guide/
+
+name:                funflow-jobs
+version:             0.1.0.0
+synopsis:            Support for distribution of flows in funflow.
+-- description:         
+homepage:            https://github.com/tweag/funflow
+license:             BSD3
+license-file:        LICENSE
+author:              Tom Nielsen
+maintainer:          nick@topos.org.uk
+-- copyright:           
+category:            Control
+build-type:          Simple
+extra-source-files:  ChangeLog.md
+cabal-version:       >=1.10
+
+library
+  exposed-modules:
+                     Control.FunFlow.Jobs.Redis
+  -- other-modules:       
+  -- other-extensions:    
+  build-depends:       base >=4.9 && <4.10
+                     , bytestring
+                     , exceptions
+                     , funflow
+                     , hedis
+                     , lens
+                     , monad-control
+                     , mtl
+                     , store
+                     , text
+                     , transformers-base
+  hs-source-dirs:      src
+  default-language:    Haskell2010
