+++ funflow/funflow.cabal
@@ -48,52 +48,49 @@ Library
                    , Control.FunFlow.Steps
                    , Control.FunFlow.Pretty
                    , Control.FunFlow.Exec.Simple
-                   , Control.FunFlow.Exec.Redis
-                   , Control.FunFlow.Exec.RedisJobs
    Other-modules:
                      Control.FunFlow.ContentStore.Notify
    Build-depends:
                  base                    >= 4.6 && <5
                , aeson                   >= 1.2.3.0
                , async
+               , bytestring
                , clock
                , constraints
-               , data-default
-               , scientific
-               , store
-               , text
-               , unordered-containers
                , containers
                , contravariant
                , cryptonite
+               , data-default
                , directory
                , exceptions
                , filepath
                , ghc-prim
                , hashable
+               , hedis
                , hostname
                , integer-gmp
                , katip                   >= 0.5.0.1
                , lens
+               , lifted-async
                , memory
+               , monad-control
                , mtl
                , path                    > 0.6.0
                , path-io
+               , pretty
                , process
                , random
-               , stm
+               , scientific
                , sqlite-simple
+               , stm
+               , store
                , template-haskell >= 2.11
+               , text
                , text
                , time
                , transformers
-               , pretty
-               , bytestring
-               , hedis
-               , monad-control
-               , transformers-base
-               , lifted-async
                , unix
+               , unordered-containers
                , vector
                , yaml
    if os(linux)
