+++ stack.yaml
@@ -38,6 +38,7 @@ resolver: lts-9.6
 packages:
 - 'funflow'
 - 'funflow-aws'
+- 'funflow-jobs'
 
 # Dependency packages to be pulled from upstream that are not in the resolver
 # (e.g., acme-missiles-0.3)
